third_party/WebKit/LayoutTests/platform/android/crypto/random-values-expected.txt

Tests crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'crypto' in self is true
PASS 'getRandomValues' in self.crypto is true
PASS self.crypto.__proto__.hasOwnProperty('getRandomValues') is true
PASS matchingBytes < 100 is true
PASS successfullyParsed is true

TEST COMPLETE
